+++ src/main/java/com/github/jlangch/venice/impl/docgen/cheatsheet/section/PrimitivesSection.java
@@ -285,6 +285,7 @@ public class PrimitivesSection implements ISectionBuilder {
         other.addSection(symbols);
         symbols.addItem(new DocItem("'a 'blue", null));
         symbols.addItem(diBuilder.getDocItem("symbol?"));
+        symbols.addItem(diBuilder.getDocItem("qualified-symbol?"));
         symbols.addItem(diBuilder.getDocItem("symbol"));
 
         final DocSection just = new DocSection("Just", "primitives.other.just");